libast: fixes for environment and locale serial number handling
This commit changes the libast ABI, but I hope to release soon-ish
with new dynamic library version numbers, so the time for such
changes is now.

src/lib/libast/include/ast_std.h,
src/lib/libast/misc/state.c,
src/lib/libast/port/astconf.c:
- Remove the redundant ast.version struct member. It is initialised
  to the value of the AST_VERSION macro and never changed, so the
  single use of it (in astconf.c, to handle 'getconf _AST_VERSION')
  can be replaced by an expansion of that macro.

src/lib/libast/include/ast_std.h,
src/lib/libast/misc/conformance.c,
src/lib/libast/misc/translate.c,
src/lib/libast/tm/tminit.c:
- ast.env_serial is an unsigned 32-bit value that is increased
  whenever an environment variable or exported shell variable is
  changed (see setenviron.c and the env_change() macro in ksh93).
  This serial number is then used to detect enviornment changes and
  trigger reinitialisation in misc/conformance.c, misc/translate.c
  and tm/tminit.c. There is a tiny chance of such reinit failing to
  happen due to wraparound if the next time the function is called
  happens exactly 2 ** 32 == 4294967296 environment changes later
  -- something that might conceivably happen if an exported
  variable is repeatedly changed in a long-running ksh shell loop.
  So, increase the size of ast.env_serial to 64 bits, reducing that
  chance to 1 in 18446744073709551616.

src/lib/libast/misc/optget.c,
src/lib/libast/misc/optlib.h:
- Backport a bugfix for ast.locale.serial handling from ksh 93v- to
  call setlocale(3) whenever the locale serial number has changed.

+++ src/lib/libast/tm/tminit.c
@@ -435,7 +435,7 @@ tmlocal(time_t now)
 void
 tminit(Tm_zone_t* zp, time_t now, const char newzone)
 {
-	static uint32_t		serial = ~(uint32_t)0;
+	static uint64_t		serial = ~(uint64_t)0;
 
 	if(tz_abbr)
 		free(tz_abbr);
